michielstokman/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.23.0
httpx>=0.25.0
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.5.0
sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
